Refactor navigation structure and update imports for improved organization
