+++ lib/pages/navigation/message_page.dart
@@ -1,0 +1,13 @@
+
+import 'package:flutter/material.dart';
+
+class MessagePage extends StatelessWidget {
+  const MessagePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('Message Page'),
+    );
+  }
+}
